key: Move error code to submodule
Move the error code out of `key/mod.rs` and into `key/error.rs`.

Code move / file creation only, no logic changes.
